refactor(knowledge_runtime): split test_config_resolver.py into smaller files
Split the 1104-line monolithic test file into 4 focused files along
class boundaries and extract shared fixtures/factories into conftest.py.
